requirements.txt
python-telegram-bot==20.7
telethon>=1.28.0
aiogram==2.25.1
python-dotenv==1.0.1
